- tmp patching keyvi to create a rust package (#136)

+++ rust/src/dictionary.rs
@@ -56,7 +56,7 @@ impl Dictionary {
         KeyviString::new(c_buf).to_owned()
     }
 
-    pub fn size(&self) -> usize {
+    pub fn size(&self) -> u64 {
         unsafe { root::keyvi_dictionary_get_size(self.dict) }
     }
 
@@ -66,7 +66,7 @@ impl Dictionary {
         KeyviMatch::new(match_ptr)
     }
 
-    pub fn get_prefix_completions(&self, key: &str, cutoff: usize) -> KeyviMatchIterator {
+    pub fn get_prefix_completions(&self, key: &str, cutoff: u64) -> KeyviMatchIterator {
         let key_c = CString::new(key).unwrap();
         let ptr = unsafe {
             root::keyvi_dictionary_get_prefix_completions(self.dict, key_c.as_ptr(), cutoff)
@@ -74,7 +74,7 @@ impl Dictionary {
         KeyviMatchIterator::new(ptr)
     }
 
-    pub fn get_fuzzy(&self, key: &str, max_edit_distance: usize) -> KeyviMatchIterator {
+    pub fn get_fuzzy(&self, key: &str, max_edit_distance: u64) -> KeyviMatchIterator {
         let key_c = CString::new(key).unwrap();
         let ptr = unsafe {
             root::keyvi_dictionary_get_fuzzy(self.dict, key_c.as_ptr(), max_edit_distance)
@@ -82,7 +82,7 @@ impl Dictionary {
         KeyviMatchIterator::new(ptr)
     }
 
-    pub fn get_multi_word_completions(&self, key: &str, cutoff: usize) -> KeyviMatchIterator {
+    pub fn get_multi_word_completions(&self, key: &str, cutoff: u64) -> KeyviMatchIterator {
         let key_c = CString::new(key).unwrap();
         let ptr = unsafe {
             root::keyvi_dictionary_get_multi_word_completions(self.dict, key_c.as_ptr(), cutoff)
